Added artifact upload back Added bundle into a single file using ncc Deleted/ignored node_modules
